Work in progress on #109. We successfully request a smaller set of racks when zooming in.
But the JS for this is kind of silly. Zooming out would already load
new racks via the BBOX strategy; zooming in just reloads a subset of
stuff we already have, when i really want to remove some from the list
instead; and I still haven't figured out how to do the important part
of modifying the HTML list of racks based on the currently visible
racks. (The callback to the protocol isn't getting fired.)

Maybe I need a custom strategy??